Add description for overrun handling
Autoware provides overrun handling to limit the resource use for each component so as to isolate the resource use on the shared computing platform and, hence, enable real-time guarantee. *[Need to confirm] Due to the long tail property of computation workloads, guaranteeing worst-case execution time on real-time schedule usually over-provisions the required resources. In order to better utilize the computation resources on shared platform and isolate the resource use, the real-time framework in Autoware monitors the resources use to handle and prevent overrun when the resources are limited. Autoware also provides the profiling tool to measure the required execution time of each component, which is used by the real-time framework to reserve the computation resource and handling overrun.*